pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rashad Philizaire <rashadphilizaire@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.2"}
pydantic = "^2.7.0"
requests = "^2.31.0"
httpx = "^0.27.0"
starlette = "^0.37.2"
sse-starlette = "^2.1.0"
tavily-python = "^0.3.3"
load-dotenv = "^0.1.0"
llama-index = "^0.10.33"
llama-index-llms-groq = "^0.1.3"
instructor = "^1.2.3"
cohere = "^5.3.4"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
logfire = {extras = ["fastapi"], version = "^0.32.0"}
llama-index-llms-azure-openai = "^0.1.8"
llama-index-llms-openai = "^0.1.19"
groq = "^0.5.0"
slowapi = "^0.1.9"
redis = "^5.0.4"
llama-index-llms-ollama = "^0.1.3"
llama-index-llms-litellm = "^0.1.4"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
click = "^8.1.7"
inquirer = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.4"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
